pyproject.toml
[project]
name = "loa-mcp-uv"
version = "0.1.0"
description = "LOA demo: Mock Workday API + MCP SSE server + agent client (uv)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langchain>=0.2.0",
  "langchain-openai>=0.1.0",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "httpx>=0.27",
  "aiosqlite>=0.20",
  "python-dotenv>=1.0",
  "fastmcp>=0.10",                 # FastMCP server with SSE transport
  "langchain-mcp-adapters>=0.1.0", # MultiServerMCPClient
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
